tests:Add RenderPassDepthStencilLayoutTransition
RenderPassDepthStencilLayoutTransition is a positive test that verifies
that both the stencil and depth aspects of a depth/stencil attachment
to a renderpass are correctly transitioned in the validation layers to
their initialLayout.
Previously there was a bug related to this functionality where only the
active aspectMask for the attachment would have its layout transitioned
which could then lead to invalid errors downstream.