scripts:Update spec.py to handle html
The spec is no longer well-formed XML so switching to lxml to handle
missing end tag cases. There were still a few tags that broke the
parsing that I had to fix manually. Those were un-closed <br> and <col>
tags.

This also modifies the parsing loop to handle further spec changes.
There is one new "hack" that was useful for the current spec but may
need to be modified/removed for future specs. That hack was to check
the error messages for a spec section that is 1 value less than the
current spec section. This helps find previous unique ids when the
section numbers get rolled forward by 1 and saves a ton of manual
re-mapping.